# String.StartsWith ordinal optimization