@jkotas 2015-09-27
> separate if statements regressed multi-character matches by 1%

A few percent is noise for this kind of micro benchmark. It will vary depending on the exact version of the processor, where the code and data happened to be exactly located in memory, ... . Could you please make the code a bit more readable - by having at least one separate if statement?