@jkotas 2015-09-27
LGTM otherwise. I like the performance data you have included in your PR description - it is exactly how performance optimization PR should look like.