@bbowyersmyth 2015-09-27
Thanks @jkotas, split it up into a separate if.