@tarekgh 2015-09-28
LGTM but I expect the case "abc".StartsWith("aX", ....) should be slower after this change as we are adding 2 extra conditions. am I missing something with this case? how we get this case 4% better?