@jkotas 2015-09-28
@bbowyersmyth could you please update the "design note" part of the commit description ?