@bbowyersmyth 2015-09-29
@jkotas Removed.
@tarekgh The original commit is not showing up now but that was with the nested conditional operators. I've rerun the tests with the current split if statement and "aX" and "ab" are -0.2%. There might be an extra advantage with conditional operators or as @jkotas suggested, noise. But the results were always pretty consistent.